new option endDateFallback
When in range-selection mode, allow submit if only start is set and
auto-complete the range selection with the fallback date.

This is useful if the available date range is known and limited, or
a sane default limit can be set. Especially on mobile devices it is
hard to communicate that a range needs to be selected.